Share the drive helper, and stop reading the drain event as success
Two review findings on the pump state machine.

`drive` was defined identically in the proptest harness and the Kani
proofs. Since it exists to pin *whether the write ran* — the observable
`advance`'s precondition is about, and one the resulting state cannot
show — two copies is exactly the arrangement that lets one harness drift
into testing something else after a signature change. Move it into
`pump_machine.rs` behind `#[cfg(any(test, kani))]`, the pattern the
guide already prescribes for helpers both need. Re-ran Kani: 11
harnesses, all verified.

The users' guide said the drain event records that the pump "drained the
rest of the input ... the pump returns success". It is emitted inside
the loop the moment the writer latches closed, before the drain
finishes, so a later fatal read still propagates on a run that already
emitted it. Say that it records draining *beginning*, and that success
depends on the drain reaching EOF.
